trading_bot/requirements.txt
requests
pandas
numpy
psycopg2-binary
ib_insync
matplotlib
scipy
pytest
jupyter
python-dotenv
